core: Keep filename intact for --out, move old meaning to --prefix
The --out flag used to be a prefix that was then being incremented to
allow downloading multiple sequences without naming clashes. However,
this resulted in some confusion for users intending to specify a single
output file (e.g. see issue #3).

Move the old logic to a new --prefix flag, and add a new --out flag that
keeps the filename. When using --out with multiple accession IDs, all
the downloaded contents are concatenated into the specified file.

Signed-off-by: Kai Blin <kblin@biosustain.dtu.dk>